CH15/ch15/langserveapp/pyproject.toml
[tool.poetry]
name = "langserveapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = "0.3.0"}
pydantic = ">2"
langchain-openai = "0.2.0"
langchain-community = "0.3.0"
qdrant-client = "^1.10.1"
python-dotenv = "^1.0.1"



[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
